requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
PyPDF2
python-docx
